merge from main

+++ BackEnd/loginmicroservices/src/main/java/com/rmit/sept/bk_loginservices/web/BookController.java
@@ -30,23 +30,39 @@ public class BookController {
     @Autowired
     private MapValidationErrorService mapValidationErrorService;
 
+    // get all the books in the catalogue
     @GetMapping(path = "/all")
     public Iterable<Book> getAllBooks() {
         return bookService.findAllBooks();
     }
 
+    // get a book from the catalogue with a specific id
     @GetMapping(path = "/{bookId}")
     public ResponseEntity<?> getBookById(@PathVariable Long bookId) {
         Book book = bookService.findById(bookId);
-        return new ResponseEntity<Book>(book, HttpStatus.OK);
+
+        if (book != null) {
+            return new ResponseEntity<Book>(book, HttpStatus.OK);
+        } else {
+            return new ResponseEntity<String>("Book with ID " + bookId + " was not found", HttpStatus.ACCEPTED);
+        }
     }
 
+    // delete a book in the catalogue with a specific id
     @DeleteMapping(path = "/{bookId}")
     public ResponseEntity<?> deleteBook(@PathVariable Long bookId) {
-        bookService.deleteBookById(bookId);
-        return new ResponseEntity<String>("Book with ID " + bookId + " was deleted", HttpStatus.OK);
+        Book book = bookService.findById(bookId);
+
+        if (book != null) {
+            bookService.deleteBookById(bookId);
+            return new ResponseEntity<String>("Book with ID " + bookId + " was deleted", HttpStatus.OK);
+        } else {
+            return new ResponseEntity<String>("Book with ID " + bookId + " was not found", HttpStatus.ACCEPTED);
+        }
+
     }
 
+    // add a new book to the catalogue
     @PostMapping("/new")
     public ResponseEntity<?> addNewBook(@Valid @RequestBody Book book, BindingResult result) {
         ResponseEntity<?> errorMap = mapValidationErrorService.MapValidationService(result);
@@ -55,41 +71,10 @@ public class BookController {
 
         Book newBook = bookService.saveBook(book);
         if (newBook != null) {
-            return new ResponseEntity<Book>(newBook, HttpStatus.CREATED);
+            return new ResponseEntity<Book>(newBook, HttpStatus.ACCEPTED);
         } else {
-            return new ResponseEntity<String>("Unable to save details for book, a copy of the book already exists.",
-                    HttpStatus.CONFLICT);
+            return new ResponseEntity<String>("Unable to add the new book, a copy of the book already exists.",
+                    HttpStatus.ACCEPTED);
         }
-    }
-
-    @GetMapping("/search/title/{title}")
-    public Iterable<Book> searchByTitle(@PathVariable String title) {
-        return bookService.getAllByTitle(title.toLowerCase());
-    }
-
-    @GetMapping("/search/authorFirstName/{authorFirstName}")
-    public Iterable<Book> searchByAuthorFirstName(@PathVariable String authorFirstName) {
-        return bookService.getAllByAuthorFirstName(authorFirstName.toLowerCase());
-    }
-
-    @GetMapping("/search/authorLastName/{authorLastName}")
-    public Iterable<Book> searchByAuthorLastName(@PathVariable String authorLastName) {
-        return bookService.getAllByAuthorLastName(authorLastName.toLowerCase());
-    }
-
-    @GetMapping("/search/sellerId/{sellerId}")
-    public Iterable<Book> searchBySellerId(@PathVariable Long sellerId) {
-        System.out.println("asdasdasdasdasdasd" + sellerId);
-        return bookService.getAllBySellerId(sellerId);
-    }
-
-    @GetMapping("/search/category/{category}")
-    public Iterable<Book> searchByCategory(@PathVariable String category) {
-        return bookService.getAllByCategory(category.toLowerCase());
-    }
-
-    @GetMapping("/search/isbn/{isbn}")
-    public Iterable<Book> searchByISBN(@PathVariable int isbn) {
-        return bookService.getAllByISBN(isbn);
     }
 }
