+++ BackEnd/loginmicroservices/src/main/java/com/rmit/sept/bk_loginservices/web/UserController.java
@@ -43,6 +43,7 @@ public class UserController {
     @Autowired
     private UserValidator userValidator;
 
+    // register a new user account
     @PostMapping("/register")
     public ResponseEntity<?> registerUser(@Valid @RequestBody User user, BindingResult result) {
         // Validate passwords match
@@ -63,6 +64,7 @@ public class UserController {
     @Autowired
     private AuthenticationManager authenticationManager;
 
+    // login using a username and password
     @PostMapping("/login")
     public ResponseEntity<?> authenticateUser(@Valid @RequestBody LoginRequest loginRequest, BindingResult result) {
         ResponseEntity<?> errorMap = mapValidationErrorService.MapValidationService(result);
@@ -78,20 +80,36 @@ public class UserController {
         return ResponseEntity.ok(new JWTLoginSucessReponse(true, jwt));
     }
 
+    // retrieve all users
     @GetMapping("/all")
     public Iterable<User> getAllUsers() {
         return userService.findAllUsers();
     }
 
+    // retrieve a user with a specific id
     @GetMapping("/{userId}")
     public ResponseEntity<?> getUserById(@PathVariable Long userId) {
         User user = userService.findById(userId);
-        return new ResponseEntity<User>(user, HttpStatus.OK);
+        if (user != null) {
+            return new ResponseEntity<User>(user, HttpStatus.OK);
+        } else {
+            return new ResponseEntity<String>("User with ID " + userId + " was not found", HttpStatus.ACCEPTED);
+        }
+
     }
 
+    // delete a user with a specific id
     @DeleteMapping("/{userId}")
     public ResponseEntity<?> deleteUser(@PathVariable Long userId) {
-        userService.deleteUserById(userId);
-        return new ResponseEntity<String>("User with ID " + userId + " was deleted", HttpStatus.OK);
+        User user = userService.findById(userId);
+
+        if (user != null) {
+            userService.deleteUserById(userId);
+            return new ResponseEntity<String>("User with ID " + userId + " was deleted", HttpStatus.OK);
+        } else {
+            return new ResponseEntity<String>("User with ID " + userId + " was not found", HttpStatus.ACCEPTED);
+        }
+
     }
+
 }
